refactor(pdf): restructure elide-pdf around concern-owned entities
Mirror the elide-office pattern: a neutral container plus small
self-addressing value entities, with each type living in the module
whose concern owns it rather than in a layer-named bag.

- document/: de-god the Pdf container into Store (data holder: object
  graph + source + limits) and a thin Pdf orchestrator whose methods
  delegate to per-capability functions over &Store; add the ObjectRole
  seam guarding structure objects from redaction.
- text/: lift the content-stream walker out of redact/ into a first-
  class text-extraction engine (walker/glyphs/cmap) that owns its
  addressing vocabulary (Address/StreamTarget, OffsetMap/OffsetRun/
  GlyphBytes, TextBlock/FontEntry, the latter now pub(crate)).
- extract/: own the extraction outputs (Embedding/EmbeddingKind/ImageId,
  Issue/IssueKind).
- redact/: own the redaction input (Detection); the shared object-graph
  guard moves to graph.rs.
- Dissolve the model/ bag entirely; unify the two Detection and two
  glyph representations into one char-offset Detection resolved through
  OffsetMap for both the text and raster paths.
- Wire the decompression-bomb bound (max_page_bytes) through the walker
  after dropping page_texts, and fold page_texts into extract().

Public API stays flat via crate-root re-exports; codec and tests import
from the concrete module paths. Behavior-preserving: the full hardening
suite (glyph-position, ToUnicode, OCG, Form-XObject recursion, quote-ops,
reflatten, flatten-verify) stays green in every feature state, cargo doc
is clean in both feature states.

+++ crates/elide-pdf/src/render/geometry.rs
@@ -94,15 +94,16 @@ pub enum GlyphSource {
 
 /// One glyph: the span of page text it covers and its box in rendered pixels.
 ///
-/// `start`/`end` are UTF-16 code-unit offsets into the page's text, so a
-/// detected span (also in UTF-16) selects the glyphs to redact.
+/// `start`/`end` are Unicode scalar (`char`) offsets into the page's text, the
+/// same unit a [`Detection`](crate::Detection) carries, so a detected span
+/// selects the glyphs to redact.
 #[derive(Debug, Clone, Copy, PartialEq)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct Glyph {
-    /// Start UTF-16 code-unit offset into the page text (inclusive).
-    pub start: u32,
-    /// End UTF-16 code-unit offset into the page text (exclusive).
-    pub end: u32,
+    /// Start character offset into the page text (inclusive).
+    pub start: usize,
+    /// End character offset into the page text (exclusive).
+    pub end: usize,
     /// The glyph's box in rendered-page pixels.
     pub rect: PixelRect,
     /// Where the geometry came from.
@@ -113,7 +114,7 @@ pub struct Glyph {
 ///
 /// This is the observation the raster redaction consumes: `pixels` is the RGB8
 /// image to overwrite, `text` is what detection ran over, and `glyphs` maps
-/// detected UTF-16 spans back to pixel rectangles.
+/// detected character spans back to pixel rectangles.
 #[derive(Debug, Clone, PartialEq)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct PageObservation {
@@ -123,7 +124,7 @@ pub struct PageObservation {
     pub width: u32,
     /// Rendered height in pixels.
     pub height: u32,
-    /// The page's text (the string `start`/`end` offsets index, in UTF-16).
+    /// The page's text (the string `start`/`end` offsets index, in characters).
     pub text: String,
     /// Every glyph's span and pixel box, in reading order.
     pub glyphs: Vec<Glyph>,
